requirements.txt
asgiref==3.5.2
autopep8==1.7.0
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
Django==4.0.5
django-filter==22.1
django-heroku==0.3.1
django-on-heroku==1.1.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycodestyle==2.9.1
PyJWT==2.6.0
pytz==2022.5
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
whitenoise==6.2.0
